Rename all Vue component files with PascalCase [DEV-43]
This is the suggested Vue style, I think it looks better, and my editor
(and/or others) might have better support for it.
