+++ app/javascript/check_ins/components/EmojiButton.vue
