+++ app/javascript/components/charts/BarGraph.vue
