+++ app/javascript/components/charts/LineChart.vue
