+++ app/javascript/groceries/components/CheckInItemsList.vue
