+++ app/javascript/groceries/components/CheckInStoreList.vue
