+++ app/javascript/groceries/components/Item.vue
