+++ app/javascript/groceries/components/LoggedInHeader.vue
