+++ app/javascript/groceries/components/StoreListEntry.vue
