+++ app/javascript/home/components/NavLink.vue
