+++ app/javascript/home/components/ParallaxImage.vue
