+++ app/javascript/logs/components/EditableTextLogRow.vue
