+++ app/javascript/logs/components/LogSelector.vue
