+++ app/javascript/logs/components/LogsIndex.vue
@@ -13,7 +13,7 @@ import { mapState } from 'pinia';
 
 import { useLogsStore } from '@/logs/store';
 
-import NewLogForm from './new_log_form.vue';
+import NewLogForm from './NewLogForm.vue';
 
 export default {
   components: {
