+++ app/javascript/logs/components/NewLogEntryForm.vue
