+++ app/javascript/logs/components/NewLogForm.vue
