+++ app/javascript/logs/components/data_renderers/CounterBarGraph.vue
@@ -9,7 +9,7 @@
 <script lang="ts">
 import { PropType } from 'vue';
 
-import BarGraph from '@/components/charts/bar_graph.vue';
+import BarGraph from '@/components/charts/BarGraph.vue';
 import { LogEntry } from '@/logs/types';
 
 export default {
