+++ app/javascript/logs/components/data_renderers/DurationTimeseries.vue
@@ -11,7 +11,7 @@
 import { CoreScaleOptions, Scale, TooltipItem } from 'chart.js';
 import { PropType } from 'vue';
 
-import LineChart from '@/components/charts/line_chart.vue';
+import LineChart from '@/components/charts/LineChart.vue';
 import { LogEntry } from '@/logs/types';
 
 function epochMsToHhMmSs(epochMs: number) {
