+++ app/javascript/logs/components/data_renderers/IntegerTimeseries.vue
@@ -9,7 +9,7 @@
 <script lang="ts">
 import { PropType } from 'vue';
 
-import LineChart from '@/components/charts/line_chart.vue';
+import LineChart from '@/components/charts/LineChart.vue';
 import { LogEntry } from '@/logs/types';
 
 export default {
