+++ app/javascript/packs/check_ins.ts
@@ -1,7 +1,7 @@
 import { ElPopover } from 'element-plus';
 import { createPinia } from 'pinia';
 
-import CheckIns from '@/check_ins/app.vue';
+import CheckIns from '@/check_ins/App.vue';
 import { renderApp } from '@/shared/customized_vue';
 
 const app = renderApp(CheckIns, '#check_ins_app');
