+++ app/javascript/packs/workout_app.ts
@@ -1,9 +1,9 @@
 import { createPinia } from 'pinia';
 
-import Modal from '@/components/modal.vue';
+import Modal from '@/components/Modal.vue';
 import { renderApp } from '@/shared/customized_vue';
 import { useElementPlus } from '@/shared/element_plus';
-import WorkoutApp from '@/workouts/workout.vue';
+import WorkoutApp from '@/workouts/Workout.vue';
 
 const app = renderApp(WorkoutApp);
 const pinia = createPinia();
