+++ app/javascript/workouts/ConfirmWorkoutModal.vue
