+++ app/javascript/workouts/WorkoutsTable.vue
